physics: Reorganize to remove private; allow recreating world
