+++ projects/physics/box2d/physics.vcxproj
@@ -62,12 +62,12 @@
     <ClInclude Include="src\ds\physics\collision.h" />
     <ClInclude Include="src\ds\physics\sprite_body.h" />
     <ClInclude Include="src\ds\physics\sprite_world.h" />
-    <ClInclude Include="src\private\contact_key.h" />
-    <ClInclude Include="src\private\contact_listener.h" />
-    <ClInclude Include="src\private\debug_draw.h" />
-    <ClInclude Include="src\private\service.h" />
-    <ClInclude Include="src\private\touch.h" />
-    <ClInclude Include="src\private\world.h" />
+    <ClInclude Include="src\ds\physics\contact_key.h" />
+    <ClInclude Include="src\ds\physics\contact_listener.h" />
+    <ClInclude Include="src\ds\physics\debug_draw.h" />
+    <ClInclude Include="src\ds\physics\service.h" />
+    <ClInclude Include="src\ds\physics\touch.h" />
+    <ClInclude Include="src\ds\physics\world.h" />
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="lib\Box2D\Collision\b2BroadPhase.cpp" />
@@ -120,12 +120,12 @@
     <ClCompile Include="src\ds\physics\collision.cpp" />
     <ClCompile Include="src\ds\physics\sprite_body.cpp" />
     <ClCompile Include="src\ds\physics\sprite_world.cpp" />
-    <ClCompile Include="src\private\contact_key.cpp" />
-    <ClCompile Include="src\private\contact_listener.cpp" />
-    <ClCompile Include="src\private\debug_draw.cpp" />
-    <ClCompile Include="src\private\service.cpp" />
-    <ClCompile Include="src\private\touch.cpp" />
-    <ClCompile Include="src\private\world.cpp" />
+    <ClCompile Include="src\ds\physics\contact_key.cpp" />
+    <ClCompile Include="src\ds\physics\contact_listener.cpp" />
+    <ClCompile Include="src\ds\physics\debug_draw.cpp" />
+    <ClCompile Include="src\ds\physics\service.cpp" />
+    <ClCompile Include="src\ds\physics\touch.cpp" />
+    <ClCompile Include="src\ds\physics\world.cpp" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <ProjectGuid>{B6E770B8-6ECC-478B-AB10-F6B01F65B433}</ProjectGuid>
